+++ aiu/main.py
@@ -498,13 +498,14 @@ def main(  # pylint: disable=R0912,R0913,R0915,R0917,R1260  # noqa: PLR0912,PLR0
         progress_display = force_progress or (LOGGER.isEnabledFor(logging.INFO) and not no_progress)
         albums = get_artist_albums(link, throw=False)
         if albums:
+            # remove duplicates that should have been skipped for real albums, but can exist for singles pseudo albums
+            album_names = sorted({album["name"] for album in albums})
             if dry:
                 LOGGER.info("Would attempt processing each album link iteratively:\n%s",
-                            json.dumps([album_info["link"] for album_info in albums], indent=2))
+                            json.dumps(album_names, indent=2))
             else:
                 # pass down all parameters except links defined by each album
-                LOGGER.info("Found albums to process:\n%s",
-                            json.dumps([album_info["name"] for album_info in albums], indent=2))
+                LOGGER.info("Found albums to process:\n%s", json.dumps(album_names, indent=2))
                 album_results = multi_fetch_albums(
                     albums,
                     # file/parsing options
@@ -512,8 +513,10 @@ def main(  # pylint: disable=R0912,R0913,R0915,R0917,R1260  # noqa: PLR0912,PLR0
                     output_file=output_file, output_dir=output_dir, output_mode=output_mode, parser_mode=parser_mode,
                     exceptions_rename_config=exceptions_rename_config, stopwords_rename_config=stopwords_rename_config,
                     # specific meta fields
-                    artist=artist, album=album, album_artist=album_artist, title=title, track=track,
-                    genre=genre, duration=duration, year=year, match_artist=match_artist,
+                    artist=artist, album_artist=album_artist,
+                    genre=genre, year=year, match_artist=match_artist,
+                    # override values that wouldn't make sense to be the same for multiple albums
+                    album=None, track=None, title=None, duration=None,
                     # heuristics flags
                     heuristic_delete_duplicates=heuristic_delete_duplicates,
                     heuristic_tag_match=heuristic_tag_match,